Fix some things I missed in the merge (kind attribute, changes to selecting audio and video tracks).

+++ Source/WebCore/html/track/AudioTrackList.cpp
@@ -76,7 +76,7 @@ void AudioTrackList::append(PassRefPtr<AudioTrack> prpTrack)
     //size_t index = track->trackIndex();
     m_tracks.append(track); // TODO: Use order correctly
 
-    printf("AudioTrackList::append(%d) len: %d\n", (int)track->trackIndex(), m_tracks.size());
+    printf("AudioTrackList::append(%d) len: %lu\n", track->trackIndex(), m_tracks.size());
     ASSERT(!track->mediaElement() || track->mediaElement() == m_owner);
     track->setMediaElement(m_owner);
